+++ src/extension.ts
@@ -2,12 +2,11 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from 'vscode';
 
-import { AdbFS } from './adbfilesystemprovider'
+import { AdbFS } from './adbfilesystemprovider';
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 export function activate(context: vscode.ExtensionContext) {
-
     // Use the console to output diagnostic information (console.log) and errors (console.error)
     // This line of code will only be executed once when your extension is activated
     console.log('"adb-filesystem" become active.');
@@ -16,6 +15,14 @@ export function activate(context: vscode.ExtensionContext) {
     context.subscriptions.push(vscode.workspace.registerFileSystemProvider('adbfs', adbfs, { isCaseSensitive: true }));
     context.subscriptions.push(vscode.commands.registerCommand('adbfs.workspaceInit', _ => {
         vscode.workspace.updateWorkspaceFolders(0, 0, { uri: vscode.Uri.parse('adbfs:/'), name: "Android Device Files" });
+    }));
+
+    // refresh file tree when the setting changed.
+    context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(e => {
+        if (e.affectsConfiguration('adbfs.sdcardFolderOnlyMode')) {
+            console.log("onDidChangeConfiguration called.");
+            vscode.commands.executeCommand("workbench.files.action.refreshFilesExplorer")
+        }
     }));
 }